fix(test): apply same fix to cli_daemon_otlp via shared common helper
cli_daemon_otlp had the same flake shape as cli_daemon_http: a
duplicated DaemonProcess that did not drain stdout, stopped reading
stderr after the handshake, accepted "Sink started" as proof the
HTTP listener was ready, and slept for a fixed 500ms before reading
counters. The protobuf and gzip-protobuf tests in particular failed
on macOS with ConnectionRefused.

Move the fixed DaemonProcess implementation, ChildGuard, http_get,
http_post, poll_until, and rsigma_bin helpers from cli_daemon_http.rs
into tests/common/mod.rs as `pub` items so both files share one
correct implementation. Add a `DaemonProcess::spawn_http` convenience
that issues the standard `daemon -r RULE --input http --api-addr
127.0.0.1:0` command line used by both test files.

Slim cli_daemon_http.rs down to its tests against the shared helper
(no behavior change there).

Update cli_daemon_otlp.rs to use the shared helper and replace the
three fixed `std::thread::sleep(Duration::from_millis(...))` calls
with `poll_until` against the specific observable conditions:
- detection_matches >= 1 for the three "triggers detection" tests
  (via a local `wait_for_detection_match` helper),
- the expected metric labels present for `otlp_metrics_exposed_after_request`.

cli_daemon_nats.rs and cli_daemon_dynamic.rs still carry their own
DaemonProcess variants (one waits on a custom log marker over NATS,
the other captures the full stderr stream for later assertions);
deferring those to a follow-up.

Verified locally on macOS:
- cargo fmt --check / clippy -D warnings clean.
- cli_daemon_http: 5 consecutive runs, 10 passed at ~1.0s each.
- cli_daemon_otlp: 5 consecutive runs, 9 passed at ~1.0s each.
- cargo test --workspace --all-features green (NATS testcontainer
  tests are flaky under concurrent execution, but pass cleanly in
  isolation; unrelated to this change).
